+++ internal/configs/mock_provider.go
@@ -32,7 +32,8 @@ func decodeMockProviderBlock(block *hcl.Block) (*Provider, hcl.Diagnostics) {
 		NameRange: block.LabelRanges[0],
 		DeclRange: block.DefRange,
 
-		Config: config,
+		// Mock providers shouldn't need any additional data.
+		Config: hcl.EmptyBody(),
 
 		// Mark this provider as being mocked.
 		Mock: true,
@@ -53,7 +54,7 @@ func decodeMockProviderBlock(block *hcl.Block) (*Provider, hcl.Diagnostics) {
 	}
 
 	var dataDiags hcl.Diagnostics
-	provider.MockData, dataDiags = decodeMockDataBody(config)
+	provider.MockData, dataDiags = decodeMockDataBody(config, MockProviderOverrideSource)
 	diags = append(diags, dataDiags...)
 
 	if attr, exists := content.Attributes["source"]; exists {
@@ -154,6 +155,16 @@ type MockResource struct {
 	DefaultsRange hcl.Range
 }
 
+type OverrideSource int
+
+const (
+	UnknownOverrideSource OverrideSource = iota
+	RunBlockOverrideSource
+	TestFileOverrideSource
+	MockProviderOverrideSource
+	MockDataFileOverrideSource
+)
+
 // Override targets a specific module, resource or data source with a set of
 // replacement values that should be used in place of whatever the underlying
 // provider would normally do.
@@ -161,13 +172,16 @@ type Override struct {
 	Target *addrs.Target
 	Values cty.Value
 
+	// Source tells us where this Override was defined.
+	Source OverrideSource
+
 	Range       hcl.Range
 	TypeRange   hcl.Range
 	TargetRange hcl.Range
 	ValuesRange hcl.Range
 }
 
-func decodeMockDataBody(body hcl.Body) (*MockData, hcl.Diagnostics) {
+func decodeMockDataBody(body hcl.Body, source OverrideSource) (*MockData, hcl.Diagnostics) {
 	var diags hcl.Diagnostics
 
 	content, contentDiags := body.Content(mockDataSchema)
@@ -212,7 +226,7 @@ func decodeMockDataBody(body hcl.Body) (*MockData, hcl.Diagnostics) {
 				}
 			}
 		case "override_resource":
-			override, overrideDiags := decodeOverrideResourceBlock(block)
+			override, overrideDiags := decodeOverrideResourceBlock(block, source)
 			diags = append(diags, overrideDiags...)
 
 			if override != nil && override.Target != nil {
@@ -229,7 +243,7 @@ func decodeMockDataBody(body hcl.Body) (*MockData, hcl.Diagnostics) {
 				data.Overrides.Put(subject, override)
 			}
 		case "override_data":
-			override, overrideDiags := decodeOverrideDataBlock(block)
+			override, overrideDiags := decodeOverrideDataBlock(block, source)
 			diags = append(diags, overrideDiags...)
 
 			if override != nil && override.Target != nil {
@@ -284,8 +298,8 @@ func decodeMockResourceBlock(block *hcl.Block) (*MockResource, hcl.Diagnostics) 
 	return resource, diags
 }
 
-func decodeOverrideModuleBlock(block *hcl.Block) (*Override, hcl.Diagnostics) {
-	override, diags := decodeOverrideBlock(block, "outputs", "override_module")
+func decodeOverrideModuleBlock(block *hcl.Block, source OverrideSource) (*Override, hcl.Diagnostics) {
+	override, diags := decodeOverrideBlock(block, "outputs", "override_module", source)
 
 	if override.Target != nil {
 		switch override.Target.Subject.AddrType() {
@@ -305,8 +319,8 @@ func decodeOverrideModuleBlock(block *hcl.Block) (*Override, hcl.Diagnostics) {
 	return override, diags
 }
 
-func decodeOverrideResourceBlock(block *hcl.Block) (*Override, hcl.Diagnostics) {
-	override, diags := decodeOverrideBlock(block, "values", "override_resource")
+func decodeOverrideResourceBlock(block *hcl.Block, source OverrideSource) (*Override, hcl.Diagnostics) {
+	override, diags := decodeOverrideBlock(block, "values", "override_resource", source)
 
 	if override.Target != nil {
 		var mode addrs.ResourceMode
@@ -342,8 +356,8 @@ func decodeOverrideResourceBlock(block *hcl.Block) (*Override, hcl.Diagnostics) 
 	return override, diags
 }
 
-func decodeOverrideDataBlock(block *hcl.Block) (*Override, hcl.Diagnostics) {
-	override, diags := decodeOverrideBlock(block, "values", "override_data")
+func decodeOverrideDataBlock(block *hcl.Block, source OverrideSource) (*Override, hcl.Diagnostics) {
+	override, diags := decodeOverrideBlock(block, "values", "override_data", source)
 
 	if override.Target != nil {
 		var mode addrs.ResourceMode
@@ -379,7 +393,7 @@ func decodeOverrideDataBlock(block *hcl.Block) (*Override, hcl.Diagnostics) {
 	return override, diags
 }
 
-func decodeOverrideBlock(block *hcl.Block, attributeName string, blockName string) (*Override, hcl.Diagnostics) {
+func decodeOverrideBlock(block *hcl.Block, attributeName string, blockName string, source OverrideSource) (*Override, hcl.Diagnostics) {
 	var diags hcl.Diagnostics
 
 	content, contentDiags := block.Body.Content(&hcl.BodySchema{
@@ -391,6 +405,7 @@ func decodeOverrideBlock(block *hcl.Block, attributeName string, blockName strin
 	diags = append(diags, contentDiags...)
 
 	override := &Override{
+		Source:    source,
 		Range:     block.DefRange,
 		TypeRange: block.TypeRange,
 	}