+++ CHANGELOG.md
@@ -10,6 +10,7 @@ UPGRADE NOTES:
     * Users of Terraform 1.6.0 and later are unaffected.
  
   This is important for users with `terraform_remote_state` data sources reading remote state across different versions of Terraform.
+* `nonsensitive` function no longer errors when applied to values that are already not sensitive. ([#33856](https://github.com/hashicorp/terraform/issues/33856))
 
 BUG FIXES:
 
@@ -19,6 +20,8 @@ BUG FIXES:
 ENHANCEMENTS:
 
 * `terraform test`: Providers defined within test files can now reference variables from their configuration that are defined within the test file. ([#34069](https://github.com/hashicorp/terraform/issues/34069))
+* `terraform test`: Providers defined within test files can now reference outputs from run blocks. ([#34118](https://github.com/hashicorp/terraform/issues/34118))
+* `terraform test`: Terraform functions are now available within variables and provider blocks within test files. ([#34204](https://github.com/hashicorp/terraform/issues/34204))
 * `import`: `for_each` can now be used to expand the `import` block to handle multiple resource instances ([#33932](https://github.com/hashicorp/terraform/issues/33932))
 
 ## Previous Releases